Allow user to configure GPU in Launch Docker image wizard in UI
