+++ src/app/pages/applications/forms/chart-form.component.ts
@@ -33,6 +33,7 @@ export class ChartFormComponent {
   protected fieldConfig: FieldConfig[];
   public fieldSets: FieldSet[] = [];
   private catalogApp: any;
+  private entityUtils = new EntityUtils();
 
   constructor(private mdDialog: MatDialog, private dialogService: DialogService,
     private modalService: ModalService, private appService: ApplicationsService) {
@@ -43,172 +44,6 @@ export class ChartFormComponent {
         this.getRow.unsubscribe();
     })
     this.utils = new CommonUtils();
-  }
-
-  createRelations(relations, parentName) {
-    const result = relations.map(relation => {
-      let relationFieldName = relation[0];
-      if (parentName) {
-        relationFieldName = `${parentName}${FORM_KEY_SEPERATOR}${relationFieldName}`;
-      }
-  
-      return {
-        action: 'SHOW',
-        when: [{
-          name: relationFieldName,
-          operator: relation[1],
-          value: relation[2],
-        }]
-      };
-    });
-
-    return result;    
-  }
-
-  parseSchemaFieldConfig(schemaConfig, parentName=null, parentIsList=false) {
-    let results = [];
-
-    if (schemaConfig.schema.hidden) {
-      return results;
-    }
-
-    let name = schemaConfig.variable;
-    if (!parentIsList && parentName) {
-      name = `${parentName}${FORM_KEY_SEPERATOR}${name}`;
-    }
-
-    let fieldConfig = {
-      required: schemaConfig.schema.required,
-      value: schemaConfig.schema.default,
-      tooltip: schemaConfig.description,
-      placeholder: schemaConfig.label,
-      name: name,
-    }
-    
-    if (schemaConfig.schema.editable === false) {
-      fieldConfig['readonly'] = true;
-    }
-
-    if (schemaConfig.schema.enum) {
-      fieldConfig['type'] = 'select';
-      fieldConfig['options'] = schemaConfig.schema.enum.map(option => {
-        return {
-          value: option.value,
-          label: option.description,
-        }
-      });
-
-    } else if (schemaConfig.schema.type == 'string') {
-      fieldConfig['type'] = 'input';
-        if (schemaConfig.schema.private) {
-          fieldConfig['inputType'] = 'password';
-          fieldConfig['togglePw'] = true;
-        }
-
-        if (schemaConfig.schema.min_length !== undefined) {
-          fieldConfig['min'] = schemaConfig.schema.min_length;
-        }
-
-        if (schemaConfig.schema.max_length !== undefined) {
-          fieldConfig['max'] = schemaConfig.schema.max_length;
-        }
-
-    } else if (schemaConfig.schema.type == 'int') {
-      fieldConfig['type'] = 'input';
-      fieldConfig['inputType'] = 'number';
-      
-    } else if (schemaConfig.schema.type == 'boolean') {
-      fieldConfig['type'] = 'checkbox';
-
-    } else if (schemaConfig.schema.type == 'hostpath') {
-      fieldConfig['type'] = 'explorer';
-      fieldConfig['explorerType'] = 'file';
-      fieldConfig['initial'] = '/mnt';
-
-    } else if (schemaConfig.schema.type == 'path') {
-      fieldConfig['type'] = 'input';
-
-    } else if (schemaConfig.schema.type == 'list') {
-
-      fieldConfig['type'] = 'list';
-      fieldConfig['label'] = `${helptext.configure} ${schemaConfig.label}`;
-      fieldConfig['width'] = '100%';
-      fieldConfig['listFields'] = [];
-
-      let listFields = [];
-      schemaConfig.schema.items.forEach(item => {
-        const fields = this.parseSchemaFieldConfig(item, null, true);
-        listFields = listFields.concat(fields);
-      });
-
-      fieldConfig['templateListField'] = listFields;
-
-    } else if (schemaConfig.schema.type == 'dict') {
-      fieldConfig = null;
-      
-      if (schemaConfig.schema.attrs.length > 0) {
-        const dictLabel = {
-          label: schemaConfig.label,
-          name: FORM_LABEL_KEY_PREFIX + name,
-          type: 'label',
-        };
-
-        if (schemaConfig.schema.show_if) {
-          dictLabel['relation'] = this.createRelations(schemaConfig.schema.show_if, parentName);
-        }
-
-        results = results.concat(dictLabel);
-      }
-
-      schemaConfig.schema.attrs.forEach(dictConfig => {
-        const subResults = this.parseSchemaFieldConfig(dictConfig, name, parentIsList);
-
-        if (schemaConfig.schema.show_if) {
-          subResults.forEach(subResult => {
-            subResult['relation'] = this.createRelations(schemaConfig.schema.show_if, parentName);
-          });
-        }
-        results = results.concat(subResults);
-      });
-    }
-
-    if (fieldConfig) {
-
-      if (fieldConfig['type']) {
-
-        if (schemaConfig.schema.show_if) {
-          fieldConfig['relation'] = this.createRelations(schemaConfig.schema.show_if, parentName);
-        }
-
-        results.push(fieldConfig);
-  
-        if (schemaConfig.schema.subquestions) {
-          schemaConfig.schema.subquestions.forEach(subquestion => {
-    
-            const subResults = this.parseSchemaFieldConfig(subquestion, parentName);
-    
-            if (schemaConfig.schema.show_subquestions_if !== undefined) {
-              subResults.forEach(subFieldConfig => {
-                subFieldConfig['isHidden'] = true;
-                subFieldConfig['relation'] = [{
-                  action: 'SHOW',
-                  when: [{
-                    name: name,
-                    value: schemaConfig.schema.show_subquestions_if,
-                  }]
-                }];
-              });
-            }
-    
-            results = results.concat(subResults);
-          });
-        }  
-      } else {
-        console.error("Unsupported type=", schemaConfig);
-      }
-    }
-
-    return results;
   }
 
   setTitle(title) {
@@ -248,7 +83,7 @@ export class ChartFormComponent {
       this.catalogApp.schema.questions.forEach(question => {
         const fieldSet = this.fieldSets.find(fieldSet => fieldSet.name == question.group);
         if (fieldSet) {
-          const fieldConfigs = this.parseSchemaFieldConfig(question);
+          const fieldConfigs = this.entityUtils.parseSchemaFieldConfig(question);
           fieldSet.config = fieldSet.config.concat(fieldConfigs);
         }
       });
@@ -258,21 +93,6 @@ export class ChartFormComponent {
     } catch(error) {
       return this.dialogService.errorReport(helptext.chartForm.parseError.title, helptext.chartForm.parseError.message);
     }
-  }
-
-  parseConfigData(configData, parentKey, result) {
-    Object.keys(configData).forEach(key => {
-      const value = configData[key];
-      let fullKey = key;
-      if (parentKey) {
-        fullKey = `${parentKey}${FORM_KEY_SEPERATOR}${key}`;
-      }
-      if (!Array.isArray(value) && (value != null && typeof value === 'object')) {
-        this.parseConfigData(value, fullKey, result);
-      } else {
-        result[fullKey] = value;
-      }
-    });
   }
 
   resourceTransformIncomingRestData(data) {
@@ -288,7 +108,7 @@ export class ChartFormComponent {
     this.parseSchema(chartSchema, true);
     this.name = data.name;
     const configData = {};
-    this.parseConfigData(data.config, null, configData);
+    this.entityUtils.parseConfigData(data.config, null, configData);
     configData['release_name'] = data.name;
     configData['changed_schema'] = true;
     
@@ -309,7 +129,7 @@ export class ChartFormComponent {
   customSubmit(data) {
     let apiCall = this.addCall;
     let values = {};
-    new EntityUtils().parseFormControlValues(data, values);
+    this.entityUtils.parseFormControlValues(data, values);
 
     let payload = [];
     payload.push({
@@ -340,9 +160,6 @@ export class ChartFormComponent {
       this.modalService.close('slide-in-form');
       this.modalService.refreshTable();
     });
-    this.dialogRef.componentInstance.failure.subscribe((err) => {
-      // new EntityUtils().handleWSError(this, err, this.dialogService);
-    })
   }
 
 }
