+++ src/app/pages/common/entity/utils.ts
@@ -265,4 +265,183 @@ export class EntityUtils {
 
     return result;
   }
+  
+  createRelations(relations:object[], parentName:string) {
+    const result = relations.map(relation => {
+      let relationFieldName = relation[0];
+      if (parentName) {
+        relationFieldName = `${parentName}${FORM_KEY_SEPERATOR}${relationFieldName}`;
+      }
+  
+      return {
+        action: 'SHOW',
+        when: [{
+          name: relationFieldName,
+          operator: relation[1],
+          value: relation[2],
+        }]
+      };
+    });
+
+    return result;    
+  }
+
+  parseSchemaFieldConfig(schemaConfig:any, parentName:string = null, parentIsList:boolean = false) {
+    let results = [];
+    let name = schemaConfig.variable;
+    if (!parentIsList && parentName) {
+      name = `${parentName}${FORM_KEY_SEPERATOR}${name}`;
+    }
+
+    let fieldConfig = {
+      required: schemaConfig.schema.required,
+      value: schemaConfig.schema.default,
+      tooltip: schemaConfig.description,
+      placeholder: schemaConfig.label,
+      name: name,
+    }
+    
+    if (schemaConfig.schema.editable === false) {
+      fieldConfig['readonly'] = true;
+    }
+
+    if (schemaConfig.schema.enum) {
+      fieldConfig['type'] = 'select';
+      fieldConfig['options'] = schemaConfig.schema.enum.map(option => {
+        return {
+          value: option.value,
+          label: option.description,
+        }
+      });
+
+    } else if (schemaConfig.schema.type == 'string') {
+      fieldConfig['type'] = 'input';
+        if (schemaConfig.schema.private) {
+          fieldConfig['inputType'] = 'password';
+          fieldConfig['togglePw'] = true;
+        }
+
+        if (schemaConfig.schema.min_length !== undefined) {
+          fieldConfig['min'] = schemaConfig.schema.min_length;
+        }
+
+        if (schemaConfig.schema.max_length !== undefined) {
+          fieldConfig['max'] = schemaConfig.schema.max_length;
+        }
+
+    } else if (schemaConfig.schema.type == 'int') {
+      fieldConfig['type'] = 'input';
+      fieldConfig['inputType'] = 'number';
+      
+    } else if (schemaConfig.schema.type == 'boolean') {
+      fieldConfig['type'] = 'checkbox';
+
+    } else if (schemaConfig.schema.type == 'hostpath') {
+      fieldConfig['type'] = 'explorer';
+      fieldConfig['explorerType'] = 'file';
+      fieldConfig['initial'] = '/mnt';
+
+    } else if (schemaConfig.schema.type == 'path') {
+      fieldConfig['type'] = 'input';
+
+    } else if (schemaConfig.schema.type == 'list') {
+
+      if (schemaConfig.schema.items.length > 0) {
+        const listLabel = {
+          label: schemaConfig.label,
+          type: 'label',
+        };
+        results = results.concat(listLabel);
+      }
+
+      fieldConfig['type'] = 'list';
+      fieldConfig['box'] = true;
+      fieldConfig['width'] = '100%';
+      fieldConfig['listFields'] = [];
+
+      let listFields = [];
+      schemaConfig.schema.items.forEach(item => {
+        const fields = this.parseSchemaFieldConfig(item, null, true);
+        listFields = listFields.concat(fields);
+      });
+
+      fieldConfig['templateListField'] = listFields;
+
+    } else if (schemaConfig.schema.type == 'dict') {
+      fieldConfig = null;
+      
+      if (schemaConfig.schema.attrs.length > 0) {
+        const dictLabel = {
+          label: schemaConfig.label,
+          name: FORM_LABEL_KEY_PREFIX + name,
+          type: 'label',
+        };
+
+        if (schemaConfig.schema.show_if) {
+          dictLabel['relation'] = this.createRelations(schemaConfig.schema.show_if, parentName);
+        }
+
+        results = results.concat(dictLabel);
+      }
+
+      schemaConfig.schema.attrs.forEach(dictConfig => {
+        const subResults = this.parseSchemaFieldConfig(dictConfig, name, parentIsList);
+
+        if (schemaConfig.schema.show_if) {
+          subResults.forEach(subResult => {
+            subResult['relation'] = this.createRelations(schemaConfig.schema.show_if, parentName);;
+          });
+        }
+        results = results.concat(subResults);
+      });
+    }
+
+    if (fieldConfig) {
+
+      if (fieldConfig['type']) {
+        results.push(fieldConfig);
+  
+        if (schemaConfig.schema.subquestions) {
+          schemaConfig.schema.subquestions.forEach(subquestion => {
+    
+            const subResults = this.parseSchemaFieldConfig(subquestion, parentName);
+    
+            if (schemaConfig.schema.show_subquestions_if !== undefined) {
+              subResults.forEach(subFieldConfig => {
+                subFieldConfig['isHidden'] = true;
+                subFieldConfig['relation'] = [{
+                  action: 'SHOW',
+                  when: [{
+                    name: name,
+                    value: schemaConfig.schema.show_subquestions_if,
+                  }]
+                }];
+              });
+            }
+    
+            results = results.concat(subResults);
+          });
+        }  
+      } else {
+        console.error("Unsupported type=", schemaConfig);
+      }
+    }
+
+    return results;
+  }
+
+  parseConfigData(configData:object, parentKey:string, result:object) {
+    Object.keys(configData).forEach(key => {
+      const value = configData[key];
+      let fullKey = key;
+      if (parentKey) {
+        fullKey = `${parentKey}${FORM_KEY_SEPERATOR}${key}`;
+      }
+      if (!Array.isArray(value) && typeof value === 'object') {
+        this.parseConfigData(value, fullKey, result);
+      } else {
+        result[fullKey] = value;
+      }
+    });
+  }
 }
